# fix(gen): use constants for enum values to prevent allocations