… examples/ex_github/oas_json_gen.go
… examples/ex_tinkoff/oas_json_gen.go