Benchstat result:
```
name                           old time/op    new time/op    delta
JSON/Sample/PetType/Decode-12    25.8ns ± 1%    13.4ns ± 1%   -48.25%  (p=0.000 n=15+14)

name                           old speed      new speed      delta
JSON/Sample/PetType/Decode-12   232MB/s ± 1%   449MB/s ± 0%   +93.33%  (p=0.000 n=15+13)

name                           old alloc/op   new alloc/op   delta
JSON/Sample/PetType/Decode-12     4.00B ± 0%     0.00B       -100.00%  (p=0.000 n=15+15)

name                           old allocs/op  new allocs/op  delta
JSON/Sample/PetType/Decode-12      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
```